# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khang Chinese Restaurant & Dimsum API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount: number;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  isAvailable: boolean;
  /** @nullable */
  ingredients?: string | null;
  /** @nullable */
  prepTime?: string | null;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  categoryId: number;
  isFeatured?: boolean;
  ingredients?: string;
  prepTime?: string;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  imageUrl?: string;
  categoryId?: number;
  isFeatured?: boolean;
  isAvailable?: boolean;
  ingredients?: string;
  prepTime?: string;
}

export interface Review {
  id: number;
  productId: number;
  reviewerName: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export interface ReviewInput {
  reviewerName: string;
  rating: number;
  comment: string;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  /** @nullable */
  imageUrl?: string | null;
}

export type OrderInputPaymentMethod =
  (typeof OrderInputPaymentMethod)[keyof typeof OrderInputPaymentMethod];

export const OrderInputPaymentMethod = {
  card: "card",
  upi: "upi",
  wallet: "wallet",
  cod: "cod",
} as const;

export type OrderInputItemsItem = {
  productId: number;
  quantity: number;
};

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  deliveryAddress: string;
  paymentMethod: OrderInputPaymentMethod;
  items: OrderInputItemsItem[];
  /** @nullable */
  notes?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  received: "received",
  preparing: "preparing",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
} as const;

export interface Order {
  id: number;
  orderCode: string;
  customerName: string;
  customerPhone: string;
  deliveryAddress: string;
  paymentMethod: string;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  tax: number;
  deliveryFee: number;
  total: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  estimatedDelivery?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  received: "received",
  preparing: "preparing",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export type OrderTrackingStatusHistoryItem = {
  status: string;
  timestamp: string;
  message: string;
};

export interface OrderTracking {
  orderCode: string;
  customerName: string;
  status: string;
  items: OrderItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
  /** @nullable */
  estimatedDelivery?: string | null;
  statusHistory: OrderTrackingStatusHistoryItem[];
}

export interface RestaurantStats {
  totalOrders: number;
  happyCustomers: number;
  menuItems: number;
  avgRating: number;
  yearsServing: number;
}

export type ListProductsParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
  /**
   * @nullable
   */
  search?: string | null;
};