# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Khang Chinese Restaurant & Dimsum API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Food products
  - name: categories
    description: Menu categories
  - name: orders
    description: Orders
  - name: reviews
    description: Product reviews
  - name: tracking
    description: Order tracking
  - name: featured
    description: Featured and top items
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products with optional category filter
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [featured]
      summary: Get featured/signature products for homepage hero
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/top:
    get:
      operationId: getTopProducts
      tags: [featured]
      summary: Get top-rated products for menu slider
      responses:
        "200":
          description: Top rated products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createProductReview
      tags: [reviews]
      summary: Add a review for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /track/{orderId}:
    get:
      operationId: trackOrder
      tags: [tracking]
      summary: Track order by order code (public)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTracking"
        "404":
          description: Order not found

  /stats:
    get:
      operationId: getStats
      tags: [featured]
      summary: Get restaurant stats (total orders, happy customers, etc.)
      responses:
        "200":
          description: Restaurant statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - productCount

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
        ingredients:
          type: ["string", "null"]
        prepTime:
          type: ["string", "null"]
      required:
        - id
        - name
        - price
        - categoryId
        - rating
        - reviewCount
        - isFeatured
        - isAvailable

    ProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        categoryId:
          type: integer
        isFeatured:
          type: boolean
        ingredients:
          type: string
        prepTime:
          type: string
      required:
        - name
        - price
        - categoryId

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        categoryId:
          type: integer
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
        ingredients:
          type: string
        prepTime:
          type: string

    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        reviewerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
      required:
        - id
        - productId
        - reviewerName
        - rating
        - comment
        - createdAt

    ReviewInput:
      type: object
      properties:
        reviewerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
      required:
        - reviewerName
        - rating
        - comment

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        imageUrl:
          type: ["string", "null"]
      required:
        - productId
        - productName
        - quantity
        - unitPrice
        - totalPrice

    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
          enum: [card, upi, wallet, cod]
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required:
              - productId
              - quantity
        notes:
          type: ["string", "null"]
      required:
        - customerName
        - customerPhone
        - deliveryAddress
        - paymentMethod
        - items

    Order:
      type: object
      properties:
        id:
          type: integer
        orderCode:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
        status:
          type: string
          enum: [received, preparing, out_for_delivery, delivered]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        tax:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderCode
        - customerName
        - customerPhone
        - deliveryAddress
        - paymentMethod
        - status
        - items
        - subtotal
        - tax
        - deliveryFee
        - total
        - createdAt
        - updatedAt

    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [received, preparing, out_for_delivery, delivered]
      required:
        - status

    OrderTracking:
      type: object
      properties:
        orderCode:
          type: string
        customerName:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        estimatedDelivery:
          type: ["string", "null"]
        statusHistory:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              timestamp:
                type: string
              message:
                type: string
            required:
              - status
              - timestamp
              - message
      required:
        - orderCode
        - customerName
        - status
        - items
        - subtotal
        - deliveryFee
        - total
        - statusHistory

    RestaurantStats:
      type: object
      properties:
        totalOrders:
          type: integer
        happyCustomers:
          type: integer
        menuItems:
          type: integer
        avgRating:
          type: number
        yearsServing:
          type: integer
      required:
        - totalOrders
        - happyCustomers
        - menuItems
        - avgRating
        - yearsServing