# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khang Chinese Restaurant & Dimsum API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from "./orderInputItemsItem";
import type { OrderInputPaymentMethod } from "./orderInputPaymentMethod";

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  deliveryAddress: string;
  paymentMethod: OrderInputPaymentMethod;
  items: OrderInputItemsItem[];
  /** @nullable */
  notes?: string | null;
}