# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khang Chinese Restaurant & Dimsum API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  /** @nullable */
  imageUrl?: string | null;
}