# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khang Chinese Restaurant & Dimsum API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  isAvailable: boolean;
  /** @nullable */
  ingredients?: string | null;
  /** @nullable */
  prepTime?: string | null;
}