# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khang Chinese Restaurant & Dimsum API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  imageUrl?: string;
  categoryId?: number;
  isFeatured?: boolean;
  isAvailable?: boolean;
  ingredients?: string;
  prepTime?: string;
}