# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khang Chinese Restaurant & Dimsum API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  productId: number;
  reviewerName: string;
  rating: number;
  comment: string;
  createdAt: string;
}